% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_genes.R
\name{available_genes}
\alias{available_genes}
\title{Identify available, matching and valid target genes from TCGA and GTEx cohorts}
\usage{
available_genes(target_genes, data)
}
\arguments{
\item{target_genes}{A character vector of target gene names}

\item{data}{A dataframe of list of dataframes.}
}
\value{
Gene or protein names available in both GTEx and TCGA cohorts for all cancers. If remove.allNA is true, also genes/proteins with all values as NA in any cancer dataframe is are removed.
}
\description{
Identify available, matching and valid target genes from TCGA and GTEx cohorts
}
\examples{
print("available_genes(genenames = ECMgenenames)")
}
