% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_nodestat.R
\name{compare_nodestats}
\alias{compare_nodestats}
\title{Compare gene-specific network statistics across cancers}
\usage{
compare_nodestats(matrigraph_summary, genes, nodestat_type = "Strength")
}
\arguments{
\item{matrigraph_summary}{The output of the matrigraph_summary function}

\item{genes}{A character vector of genenames}

\item{nodestat_type}{With the default usage of matrigraph_summary, this argument  could be "Betweenness", "Closeness", "Strength", or "ExpectedInfluence".}
}
\value{
A grouped barplot
}
\description{
Compare gene-specific network statistics across cancers
}
\examples{
print("compare_nodestats(matrigraph_summary, genes)")
}
