% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edges_2_JSDF.R
\name{edges_2_JSDF}
\alias{edges_2_JSDF}
\title{Map an edgelist to a sample-level JS dataframe}
\usage{
edges_2_JSDF(
  discreteDF = scTHPA_matridata$discrete$`adipose tissue_fibroblasts`,
  edgelist = scTHPA_matrigraph$`adipose tissue_fibroblasts`$edge_df
)
}
\arguments{
\item{discreteDF}{A single discrete matridata object}

\item{edgelist}{Matridata edgelist with two columns}
}
\value{
A dataframw with nrow(edgelist) columns and nrow(discreteDF) rows.
}
\description{
Map an edgelist to a sample-level JS dataframe
}
\examples{
NULL
}
