% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Globals.R
\name{Globals}
\alias{Globals}
\title{Global variables}
\usage{
utils::globalVariables(c("xlab", "ylab", "Data", "Value", "Genes", "theme_neighborhood"))
}
\value{
Global variables.
}
\description{
Global variables
}
\keyword{internal}
