% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrigraph_summary.R
\name{matrigraph_summary}
\alias{matrigraph_summary}
\title{Matrigraph summary by node.stat attribute}
\usage{
matrigraph_summary(output_matrigraph, metric)
}
\arguments{
\item{output_matrigraph}{The output of the matrinet_estimate function}

\item{metric}{A network weight type, i.e., one of the column names in matrigraph$edge.df}
}
\value{
A list of node summary specific dataframes where columns represents cancer specific statistics
}
\description{
Matrigraph summary by node.stat attribute
}
\examples{
print(c("matrigraph_summary"))
}
