% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrigraph_to_adjacency.R
\name{matrigraph_to_adjacency}
\alias{matrigraph_to_adjacency}
\title{Convert matrigraphs to weighted adjacency matrices}
\usage{
matrigraph_to_adjacency(output_matrigraph, weights_by = "MI_D")
}
\arguments{
\item{output_matrigraph}{The output from the matrinet_estimate function.}

\item{weights_by}{Column name in the matrigraph edge.df by which the network weigts are defined.}
}
\value{
A list of weighted adjacency matrices
}
\description{
Convert matrigraphs to weighted adjacency matrices
}
\examples{
print("matrigraph_to_adjacency(output_matrigraph = matrinet_TCGA)")
}
