% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrinet_data.R
\name{matrinet_data}
\alias{matrinet_data}
\title{Test title}
\usage{
matrinet_data(datalist, quantiles, genenams)
}
\arguments{
\item{datalist}{A dataframe or list of group (cancer) dataframes containing target gene variables as column vectors. Additional variables (e.g. sampleIDs) could be included, target genes/proteins are extracted automatically from the dataframe by the column names of prior_topology object}

\item{quantiles}{A dataframe or list of cancer-specific gene expression discretization levels for each gene.}

\item{genenams}{A character vector of target gene names}
}
\value{
A matridata object that can be used as an input for matrinet_estimate function.
}
\description{
Test title
}
\examples{
print("matrinet_data(datalist, quantiles, genenames)")
}
