% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrinet_estimate.R
\name{matrinet_estimate}
\alias{matrinet_estimate}
\title{Matrinet network estimation function}
\usage{
matrinet_estimate(matrigraph, matridata)
}
\arguments{
\item{matrigraph}{Output object from the matrinet_graph function}

\item{matridata}{Output object from the matrinet_data function}
}
\value{
Matrigraph object with added results from the estimated networks
}
\description{
Matrinet network estimation function
}
\examples{
print(c("matrinet_estimate(matrigraph, matridata)"))
}
