% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrinet_graph.R
\name{matrinet_graph}
\alias{matrinet_graph}
\title{Create matrigraph object}
\usage{
matrinet_graph(
  matridata,
  prior_topology,
  add_node_df = NULL,
  add_edge_df = NULL
)
}
\arguments{
\item{matridata}{The output from matrinet_data() function or any list of group (cancer) specific data-frames containing node covariates, e.g., gene or protein expression data. Extracted automatically from the dataframe by the column names of prior_topology object.}

\item{prior_topology}{An adjacency matrix providing a fixed skeleton for the network. The default structure is given by the matrixDB interaction data.}

\item{add_node_df}{A data-frame for custom node attributes that are common for each group (e.g. protein/gene family).}

\item{add_edge_df}{A data-frame for custom edge attributes that are common for each group (e.g. gene-gene interaction data source).}
}
\value{
Returns a list of group (cancer) specific matrigraph objects, with 1) node_df, 2) edge_df and 3) summary dataframes. Matrigraph object is meant for storing network estimation results and custom node/edge attributes.
}
\description{
Create the initial skeleton of matrigraph object for storing network data.
}
\examples{
print(c("matrinet_graph(matridata, prior_topology)"))
}
