% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrinet_init.R
\name{matrinet_init}
\alias{matrinet_init}
\title{Initializing input dataframes for the matrinetR package}
\usage{
matrinet_init(
  cancers = c("laml", "ov", "prad", "brca"),
  TCGA,
  GTEx,
  cohort = "TCGA",
  subset_genes
)
}
\arguments{
\item{cancers}{A character vector of cancer abbreviations. The corresponding datasets are extracted from the TCGA and GTEx objects.}

\item{TCGA}{A list of cancer specific TCGA datasets named by the TCGA abbreviation standards (lowercase symbols).}

\item{GTEx}{A list of cancer specific GTEx datasets named by the TCGA abbreviation standards (lowercase symbols).}

\item{cohort}{Select a cohort}

\item{subset_genes}{A character vector representing the names of genes to be extracted}
}
\value{
Initialized data object to be used as an input for the matrinet_data function.
}
\description{
Initializing input dataframes for the matrinetR package
}
\examples{
print("matrinet_init()")
}
