% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrisome_GTEx.R
\docType{data}
\name{matrisome_GTEx}
\alias{matrisome_GTEx}
\title{GTEx matrisome gene expression}
\format{
A list of tissue specific datasets each of which consisting of with 365 variables (columns):
\itemize{
  \item SampleID and (primary) Site. Xenahub dataset ID: GTEX_phenotype
  \item Normalized and log2(norm_count+1) transformed matrisome mRNA data: Xenahub ID: gtex_RSEM_Hugo_norm_count
}
}
\usage{
matrisome_GTEx
}
\description{
A list of tissue specific datasets containing matrisome gene-expression data downloaded from UCSC Xenahub platform's GTEx cohort \url{https://xenabrowser.net/datapages/}
}
\keyword{datasets}
