% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixDB_annotations.R
\docType{data}
\name{matrixDB_annotations}
\alias{matrixDB_annotations}
\title{Matrisome interaction data from matrixDB}
\format{
A data frame with 967 rows and 6 variables:
\itemize{
  \item gene: Gene name
  \item category: Gene specific category
  \item family: Gene specific family
}
}
\usage{
matrixDB_annotations
}
\description{
A dataset containing matrisome gene specific annotations
}
\keyword{datasets}
