% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighborhood_plot.R
\name{neighborhood_plot}
\alias{neighborhood_plot}
\title{Plot neighborhood curves}
\usage{
neighborhood_plot(
  matrigraph_output,
  prior_topology,
  center_gene = "SDC1",
  weights = "MI_D",
  smoothness = 0.5
)
}
\arguments{
\item{matrigraph_output}{The matrigraph output}

\item{prior_topology}{Prior network topology consisting of genes in the matrigraph_output}

\item{center_gene}{Name of the center gene}

\item{weights}{Weight type}

\item{smoothness}{Curve smoothness}
}
\value{
Neighborhood curve plot
}
\description{
Plot neighborhood curves
}
\examples{
print("1")
}
